Move directory from sponge to sponge7 and add SpongeModule
