+++ plugin/sponge/src/main/java/broccolai/tickets/sponge7/model/SpongeOnlineSoul.java
@@ -1,4 +1,4 @@
-package broccolai.tickets.sponge.model;
+package broccolai.tickets.sponge7.model;
 
 import broccolai.tickets.api.model.user.OnlineSoul;
 import org.spongepowered.api.command.CommandSource;
